DRY spec helpers
