+++ spec/features/form/check_answers_page_spec.rb
@@ -1,6 +1,8 @@
 require "rails_helper"
+require_relative "helpers"
 
 RSpec.describe "Form Check Answers Page" do
+  include Helpers
   let!(:case_log) { FactoryBot.create(:case_log, :in_progress) }
   let!(:empty_case_log) { FactoryBot.create(:case_log) }
   let(:id) { case_log.id }
@@ -11,23 +13,6 @@ RSpec.describe "Form Check Answers Page" do
 
   let(:subsection) { "household_characteristics" }
   let(:conditional_subsection) { "conditional_question" }
-
-  def fill_in_number_question(case_log_id, question, value, path)
-    visit("/case_logs/#{case_log_id}/#{path}")
-    fill_in("case-log-#{question.to_s.dasherize}-field", with: value)
-    click_button("Save and continue")
-  end
-
-  def answer_all_questions_in_income_subsection(case_log)
-    visit("/case_logs/#{case_log.id}/net_income")
-    fill_in("case-log-earnings-field", with: 18_000)
-    choose("case-log-incfreq-yearly-field")
-    click_button("Save and continue")
-    choose("case-log-benefits-all-field")
-    click_button("Save and continue")
-    choose("case-log-hb-prefer-not-to-say-field")
-    click_button("Save and continue")
-  end
 
   context "when the user needs to check their answers for a subsection" do
     it "can be visited by URL" do
